# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Smart Buy",
        "description": "Smart Buy",
        "contact": {
            "name": "Gaurav Patel",
            "email": "gaurav.gkp29@gmail.com",
            "url": "/"
        }
    },
    "paths": {
        "/auth/sign-in": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "Signin",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Signin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "Optional -> if you get"
                                    },
                                    "email": {
                                        "type": "string",
                                        "description": "Optional -> if you get"
                                    },
                                    "social_type": {
                                        "type": "string",
                                        "enum": ["google", "facebook"]
                                    },
                                    "social_id": {
                                        "type": "string",
                                        "description": "Pass only while using social"
                                    },
                                    "mobile_country_code": {
                                        "type": "string",
                                        "description": "Optional"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "description": "Optional"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "description": "Optional"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.Signin"
            }
        },
        "/auth/get-lottie-animation": {
            "get": {
                "x-controller-name": "AuthController",
                "x-operation-name": "LottieAnimation",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LottieAnimation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "AuthController.LottieAnimation"
            }
        },

        "/product/feed": {
            "get": {
                "x-controller-name": "ProductController",
                "x-operation-name": "Feed",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Feed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "page",
                    "in": "query",
                    "schema": {
                        "type": "number"
                    }
                }],
                "operationId": "ProductController.Feed"
            }
        },
        "/product/get-results": {
            "get": {
                "x-controller-name": "ProductController",
                "x-operation-name": "GetResults",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetResults",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "img_urls",
                    "in": "query",
                    "description": "Multiple images pass | seperate ex. 1|2|3 like that",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "ProductController.GetResults"
            }
        },
        "/product/get-cache-results": {
            "get": {
                "x-controller-name": "ProductController",
                "x-operation-name": "CacheResults",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CacheResults",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "Product_Url",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "ProductController.CacheResults"
            }
        },
        "/product/get-image-key": {
            "get": {
                "x-controller-name": "ProductController",
                "x-operation-name": "ImageKey",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ImageKey",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "Img_url",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "ProductController.ImageKey"
            }
        },
        "/product/get-product-key": {
            "get": {
                "x-controller-name": "ProductController",
                "x-operation-name": "ProductKey",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ProductKey",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "Product_Url",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "ProductController.ProductKey"
            }
        },
        "/product/data-insert": {
            "get": {
                "x-controller-name": "ProductController",
                "x-operation-name": "DataInserter",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "DataInserter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "Host",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "Product_Url",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "Product_Title",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "Img_url",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "SellPrice",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "ProductController.DataInserter"
            }
        },
        "/product/data-insert-post": {
            "post": {
                "x-controller-name": "ProductController",
                "x-operation-name": "DataInserterPost",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "DataInserterPost",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "array"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProductController.DataInserterPost"
            }
        },
        "/product/data-insert-android": {
            "post": {
                "x-controller-name": "ProductController",
                "x-operation-name": "DataInserterAndroid",
                "tags": ["ProductController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "DataInserterAndroid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "array"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProductController.DataInserterAndroid"
            }
        },

        "/banner/list": {
            "get": {
                "x-controller-name": "BannerController",
                "x-operation-name": "List",
                "tags": ["BannerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [],
                "operationId": "BannerController.List"
            }
        },

        "/report/category-list": {
            "get": {
                "x-controller-name": "ReportController",
                "x-operation-name": "CategoryList",
                "tags": ["ReportController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CategoryList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [],
                "operationId": "ReportController.CategoryList"
            }
        },
        "/report/user-send-report": {
            "post": {
                "x-controller-name": "ReportController",
                "x-operation-name": "UserSendReport",
                "tags": ["ReportController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UserSendReport",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "report_category_id": {
                                        "type": "string",
                                        "description": "Pass  Multiple comma seperate like 1,2,3"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ReportController.UserSendReport"
            }
        },

        "/scrap/get-background-scrap-list": {
            "get": {
                "x-controller-name": "ScrapController",
                "x-operation-name": "List",
                "tags": ["ScrapController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [],
                "operationId": "ScrapController.List"
            }
        },
        "/scrap/insert-background-scrap-url": {
            "post": {
                "x-controller-name": "ScrapController",
                "x-operation-name": "InsertUrl",
                "tags": ["ScrapController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "InsertUrl",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "description": "Pass  Multiple comma seperate like 1,2,3"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ScrapController.InsertUrl"
            }
        }
    },
    "components": {},
    "servers": [{
        "url": "http://143.110.249.142:3001/api"
    }, {
        "url": "http://localhost:3001/api"
    }]
}